# server/gen/shrls.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "shrls.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Shrls"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/shrl/{shrl.alias}": {
      "get": {
        "operationId": "Shrls_GetShrl",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/shrlsGetShrlResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "shrl.alias",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "shrl.id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Shrls"
        ]
      }
    },
    "/v1/shrls/{shrl.alias}": {
      "get": {
        "operationId": "Shrls_GetShrls",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/shrlsGetShrlsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "shrl.alias",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "shrl.id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Shrls"
        ]
      }
    }
  },
  "definitions": {
    "ShortURLShortURLType": {
      "type": "string",
      "enum": [
        "LINK",
        "UPLOAD",
        "SNIPPET"
      ],
      "default": "LINK"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "shrlsExpandedURL": {
      "type": "object",
      "properties": {
        "url": {
          "$ref": "#/definitions/shrlsRedirect"
        },
        "file": {
          "type": "string",
          "format": "byte"
        },
        "snippet": {
          "$ref": "#/definitions/shrlsSnippet"
        }
      }
    },
    "shrlsGetShrlResponse": {
      "type": "object",
      "properties": {
        "shrl": {
          "$ref": "#/definitions/shrlsShortURL"
        }
      }
    },
    "shrlsGetShrlsResponse": {
      "type": "object",
      "properties": {
        "shrl": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/shrlsShortURL"
          }
        }
      }
    },
    "shrlsRedirect": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        },
        "favicon": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "Url Redirects"
    },
    "shrlsRefShortURL": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "alias": {
          "type": "string"
        }
      },
      "title": "Reference ShortURL"
    },
    "shrlsShortURL": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "External ID for ShortURL"
        },
        "createdAt": {
          "type": "string",
          "format": "int64",
          "title": "Date the ShortURL was created at"
        },
        "type": {
          "$ref": "#/definitions/ShortURLShortURLType",
          "title": "ShortURL type"
        },
        "stub": {
          "type": "string",
          "title": "URL stub"
        },
        "content": {
          "$ref": "#/definitions/shrlsExpandedURL",
          "title": "Expanded destination"
        },
        "views": {
          "type": "string",
          "format": "int64",
          "title": "Number of times the link has been clicked"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "User defined tags for the ShortURL to aid searching"
        }
      },
      "title": "========================================\nObjects\n========================================\nShortened Urls"
    },
    "shrlsSnippet": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "body": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "Text snippets"
    }
  }
}